Adding this to github.
The initial work for Download All The Things. Created a form for entering user information. The form minimizes to the tray and the tray has a context menu for setting up a new user account or exiting the program. There is a stubbed out function for going out to the web and grabbing the actual bandwidth usage.