I have no idea what I (comcast) am (is) doing.
This perl script should go to the specified URL, enter the login credentials, and then read the usage information from the page. However, right now it seems to enter the login credentials and get stuck on a page that is loading the information. It will probably take some digging to figure out the way around this issue. Not entirely sure what comcast is doing on this page, probably something stupid.