Added an untested phantomjs script.
This was my attempt to use phantomjs to scrape the data from the website. Entirely untested because comcast chose this exact moment to upgrade their website. I may end up using some other tools to make this process easier. I don't think the form is actually being submitted at the point. This might be my mistake as my javascript is a bit rusty.